% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_idaifield.R
\name{add_limit_to_query}
\alias{add_limit_to_query}
\title{Add limit to JSON query}
\usage{
add_limit_to_query(query, conn)
}
\arguments{
\item{query}{A MongoDB JSON-query as used in this package.}

\item{conn}{A connection object returned by \code{\link[=connect_idaifield]{connect_idaifield()}}.}
}
\value{
A query with limit added for PouchDB
}
\description{
This function adds a limit of the max db docs to a query.
}
\examples{
\dontrun{
conn <- connect_idaifield(project = "test", pwd = "hallo")
fields <- c("resource.category", "resource.identifier")
query <- paste0('{ "selector": { "$not": { "resource.id": "" } },
   "fields": [', paste0('"', fields, '"', collapse = ", "), '] }')
add_limit_to_query(query, conn)
}
}
\keyword{internal}
