% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{bce_ce}
\alias{bce_ce}
\title{Translate a dating value from iDAI.field to a positive or negative number}
\usage{
bce_ce(list)
}
\arguments{
\item{list}{A named list containing at least the following elements:
\describe{
\item{inputYear}{A numerical value representing a year.}
\item{inputType}{A character string indicating the type of dating
for the inputYear value. Must be one of "bce", "ce", or "bp".}
}}
}
\value{
A numerical value representing the year, with a negative
sign indicating BCE and a positive sign indicating CE.
}
\description{
This function takes a list containing a numerical year and a type of dating
(either "bce", "ce", or "bp") and returns the year as a number with a
positive or negative sign indicating whether the year is BCE or CE.
If the dating type is "bp", the year is subtracted from 1950
to get a BCE year.
}
\examples{
\dontrun{
list <- list(inputYear = 100, inputType = "bce")
bce_ce(list)
}
}
\seealso{
\code{\link{simplify_idaifield}}
}
\keyword{internal}
