% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_conn_for_project}
\alias{check_conn_for_project}
\title{TEMPORARY FUNCTION to warn about project(name)}
\usage{
check_conn_for_project(project = connection$project, fail = FALSE)
}
\arguments{
\item{project}{connection$project}

\item{fail}{should there be an error?}
}
\value{
TRUE/FALSE
}
\description{
TEMPORARY FUNCTION to warn about project(name)
}
\keyword{internal}
