% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_for_sublist}
\alias{check_for_sublist}
\title{Checks if a list has sub-lists and returns TRUE if so}
\usage{
check_for_sublist(single_resource_field)
}
\arguments{
\item{single_resource_field}{a list to be checked}
}
\value{
TRUE/FALSE
}
\description{
Checks if a list has sub-lists and returns TRUE if so
}
\examples{
\dontrun{
list <- list(1, 2, 3, list("börek", 2, 3))

check_for_sublist(list)
}
}
\keyword{internal}
