% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_idaifieldr_version.R
\name{check_idaifieldr_version}
\alias{check_idaifieldr_version}
\title{Check Version of idaifieldR}
\usage{
check_idaifieldr_version(installed_version = getNamespaceVersion("idaifieldR"))
}
\arguments{
\item{installed_version}{version of idaifieldR currently in use}
}
\value{
TRUE if current release is used, FALSE if not.
}
\description{
Checks if a new version of idaifieldR is available (runs on attach).
}
\examples{
\dontrun{
check_idaifieldr_version(packageVersion('idaifieldR'))
}
}
