% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_idaifieldr_version.R
\name{check_idf_version}
\alias{check_idf_version}
\title{Check Version of idaifieldR}
\usage{
check_idf_version(installed_version = getNamespaceVersion("idaifieldR"))
}
\value{
nothing
}
\description{
Runs on load and checks if a new version of idaifieldR is available.
}
\examples{
\dontrun{
check_idf_version(packageVersion('idaifieldR'))
}
}
