% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_if_idaifield}
\alias{check_if_idaifield}
\title{check_if_idaifield}
\usage{
check_if_idaifield(testobject)
}
\arguments{
\item{testobject}{An object that should be evaluated.}
}
\value{
a matrix that allows other functions to determine which type of
list the object is
}
\description{
For internal use... checks if an object can actually processed by
the functions in this package which need the specific format that is
returned by the core function get_idaifield_docs(...).
}
\examples{
\dontrun{
idaifield_docs <- get_idaifield_docs(projectname = "rtest",
connection = connect_idaifield(serverip = "192.168.1.21",
user = "R",
pwd = "password"))

check_if_idaifield(idaifield_docs)
}
}
\keyword{internal}
