% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_if_uid}
\alias{check_if_uid}
\title{check_if_uid}
\usage{
check_if_uid(string)
}
\arguments{
\item{string}{A character string that should be checked for being a UID
as used in i.DAIfield 2 (expects a single character value!)}
}
\value{
TRUE if UID, or FALSE if not
}
\description{
check_if_uid
}
