% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_idaifield.R
\name{connect_idaifield}
\alias{connect_idaifield}
\title{Connect to an iDAI.field / Field Desktop Client}
\usage{
connect_idaifield(
  serverip = "localhost",
  project = NULL,
  pwd = "password",
  ping = TRUE
)
}
\arguments{
\item{serverip}{The IP address of the Field Client, or "localhost". Leave
the default if you are using Field Desktop on the same computer.}

\item{project}{(\emph{required}) The name of the project you want to work with.
For a list of available projects, see \code{\link[=idf_projects]{idf_projects()}}.}

\item{pwd}{(\emph{required}) The password used to authenticate with the Field
Client (default is "password").}

\item{ping}{Should the connection be pinges on creation? Defaults to TRUE.}
}
\value{
\code{\link[=connect_idaifield]{connect_idaifield()}} returns an \code{idf_connection_settings}
object that is used by other functions in this package to retrieve data
from the
\href{https://github.com/dainst/idai-field}{iDAI.field / Field Desktop Client}.
}
\description{
This function establishes a connection to the database of your
\href{https://github.com/dainst/idai-field}{iDAI.field / Field Desktop Client},
and returns a connection object containing the necessary information for
other functions to access the database, such as \code{\link[=get_idaifield_docs]{get_idaifield_docs()}},
\code{\link[=idf_query]{idf_query()}}, \code{\link[=idf_index_query]{idf_index_query()}}, or \code{\link[=idf_json_query]{idf_json_query()}}.
}
\details{
By default, if you are using Field Desktop on the same machine,
you should not need to specify the \code{serverip} argument, as it defaults to
"localhost". If you want to access a client on the same network that is
not running on the same computer as R, you can supply the local IP
(without the port (':3000')).
The \code{pwd} argument needs to be set to the password that
is set in your Field Desktop Client under \emph{Tools/Werkzeuge} >
\emph{Settings/Einstellungen}: 'Your password'/'Eigenes Passwort'.
\code{project} has to be set to the identifier of the project database you
will query.
}
\examples{
\dontrun{
conn <- connect_idaifield(
  serverip = "localhost",
  pwd = "hallo",
  project = "rtest"
)

conn$status

idf_ping(conn)
}
}
\seealso{
\itemize{
\item Ping the connection with \code{\link[=idf_ping]{idf_ping()}}
\item Get a list of projects in the database with \code{\link[=idf_projects]{idf_projects()}}
}
}
