% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_geometry.R
\name{convert_polygon}
\alias{convert_polygon}
\title{Converts a polygon list to a matrix}
\usage{
convert_polygon(geometry_polygon)
}
\arguments{
\item{geometry_polygon}{a list of the format that any polygon has
in the iDAI.field 2 database}
}
\value{
a matrix that displays the same polygon
}
\description{
Converts a polygon list to a matrix
}
\examples{
\dontrun{
test_2 <- list(type = "Polygon", coordinates = list(list(list(1), list(1)),
list(list(1), list(2)),
list(list(2), list(2)),
list(list(1), list(1))))

convert_polygon(test_2$coordinates)
}
}
