% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_geometry.R
\name{convert_to_coordmat}
\alias{convert_to_coordmat}
\title{Converts coordinate list from iDAI.field to a Matrix}
\usage{
convert_to_coordmat(coordinates)
}
\arguments{
\item{coordinates}{a list of the format that any coordinateslist has
in the iDAI.field 2 / Field Desktop database}
}
\value{
a matrix that displays the same coordinates
}
\description{
Converts coordinate list from iDAI.field to a Matrix
}
\examples{
\dontrun{
test_2 <- list(type = "Polygon", coordinates = list(list(list(1), list(1)),
list(list(1), list(2)),
list(list(2), list(2)),
list(list(1), list(1))))

convert_polygon(test_2$coordinates)
}
}
\keyword{internal}
