% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importlist.R
\name{convert_to_df}
\alias{convert_to_df}
\title{Tries to convert the list that one gets from the json-format into a usable df
this is currently horrible}
\usage{
convert_to_df(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{}
}
\value{

}
\description{
Tries to convert the list that one gets from the json-format into a usable df
this is currently horrible
}
