% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{convert_to_onehot}
\alias{convert_to_onehot}
\title{Break down a List from a Checkbox Field to Onehot-Coded Values}
\usage{
convert_to_onehot(resource, inputtypes, except = NULL)
}
\arguments{
\item{resource}{The resource to process (from an \code{idaifield_resources}-list).}

\item{inputtypes}{A matrix of fields with the given inputType as
returned by \code{\link[=parse_field_inputtypes]{parse_field_inputtypes()}}}

\item{except}{A vector of fieldnames that should be ignored.}
}
\value{
The resource object with the values of checkboxes
separated into one-hot-coded versions.
}
\description{
This function converts all checkboxes fields (except the ones
listed in the \code{except}-param) into "one-hot" coded list items.
}
\examples{
\dontrun{
...
}
}
\seealso{
\itemize{
\item Needs output of: \code{\link[=parse_field_inputtypes]{parse_field_inputtypes()}}
}
}
