% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_language_list.R
\name{download_language_list}
\alias{download_language_list}
\title{Download a Language-List from GitHub}
\usage{
download_language_list(project = "core", language = "en")
}
\arguments{
\item{project}{name of the project for which the language files should be
downloaded; case sensitive! Has to match the name in the Language-file
exactly. If default ("core") is used, the common language file from the core
library will be downloaded.}

\item{language}{Language short name that is to be extracted, e.g. "en",
defaults to "en"}
}
\value{
a list that can be processed with \code{\link{get_language_lookup}}.
}
\description{
This function downloads language lists from the
iDAI.field-GitHub-repository and can be used to supply
additional lists to \code{\link{get_language_lookup}}.
}
\examples{
\dontrun{
lang_list <- download_language_list(language = "de")
}
}
\references{
Field Desktop Client: \url{https://github.com/dainst/idai-field}
}
