% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_language_lookup.R
\name{extract_field_names}
\alias{extract_field_names}
\title{Create a data.frame from a list of labels and descriptions from iDAI.field}
\usage{
extract_field_names(fields_list)
}
\arguments{
\item{fields_list}{A named list that contains one or two other
named lists ("label" and "description") with the translation / display
language of the respective internal value (i.e. the name of the list)}

\item{language}{}
}
\value{
a data frame with the column "var" and "label" containing the
value in var and its respective translation / display value in "label"
}
\description{
Helper to get_language_lookup()
}
\examples{
\dontrun{
fields_list <- list("category" = list("label" = "Category"),
                    "identifier" = list("label" = "Name / ID (unique)",
                    "description" = "Description of the field"))
df <- extract_field_names(fields_list)
}
}
\keyword{internal}
