% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{extract_inputtypes}
\alias{extract_inputtypes}
\title{Extracts a List of Input Types from the Project Configuration}
\usage{
extract_inputtypes(nested_list, parent_name = NULL, category_name = NULL)
}
\arguments{
\item{nested_list}{A configuration list as returned
by \code{\link[=get_configuration]{get_configuration()}}}

\item{parent_name}{Used for recursive behaviour}

\item{category_name}{Used for recursive behaviour}
}
\value{
A list
}
\description{
Internal helper to \code{\link[=get_field_inputtypes]{get_field_inputtypes()}}
}
\examples{
\dontrun{
conn <- connect_idaifield(serverip = "localhost",
                          pwd = "hallo",
                          project = "rtest")
config <- get_configuration(connection = conn)
input_types <- extract_inputtypes(config)
}
}
\keyword{internal}
