% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_layer.R
\name{find_categories}
\alias{find_categories}
\title{returns the category of the ids (UUID/identifier)}
\usage{
find_categories(ids, index, id_type)
}
\arguments{
\item{index}{A data.frame as returned by \code{\link[=get_field_index]{get_field_index()}} or
\code{\link[=make_index]{make_index()}}.}

\item{id_type}{One of "identifier" or "UID" - Should the function use
\emph{identifier}s or \emph{UUID}s to find the layer?}
}
\value{
a vector of categories for each resource
}
\description{
returns the category of the ids (UUID/identifier)
}
\seealso{
\itemize{
\item \code{\link[=find_layer]{find_layer()}}.
}
}
\keyword{internal}
