% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_layer.R
\name{find_parents}
\alias{find_parents}
\title{returns the UUID/identifier of the resources the 'ids' are contained in}
\usage{
find_parents(ids, uidlist, id_type)
}
\arguments{
\item{uidlist}{A data.frame as returned by \code{\link[=get_field_index]{get_field_index()}} or
\code{\link[=get_uid_list]{get_uid_list()}}.}

\item{id_type}{One of "identifier" or "UID" - Should the function use
\emph{identifier}s or \emph{UUID}s to find the layer?}
}
\value{
a vector of resources in which each id is located
}
\description{
returns the UUID/identifier of the resources the 'ids' are contained in
}
\seealso{
\itemize{
\item \code{\link[=find_layer]{find_layer()}}.
}
}
\keyword{internal}
