% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_docs.R
\name{find_resource}
\alias{find_resource}
\title{Finding the "resource"-list in a CouchDB-output}
\usage{
find_resource(list)
}
\arguments{
\item{list}{a list formatted from a CouchDB-JSON-output}
}
\value{
the list of resource-lists
}
\description{
Finding the "resource"-list in a CouchDB-output
}
\examples{
\dontrun{
testlist <- list(docs = list(list(resource = list(test = "test",
                                                  test2 = "test2")),
                             list(resource = list(test = "test",
                                                  test2 = "test2"))),
                 warning = "warning")
find_resource(testlist)
}
}
\keyword{internal}
