% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{gather_languages}
\alias{gather_languages}
\title{Gather Fields with Multiple Language Values}
\usage{
gather_languages(input_list, language = "all", silent = FALSE)
}
\arguments{
\item{input_list}{A list with character values containing (or not)
sublists for each language.}

\item{language}{The short name (e.g. "en", "de", "fr") of the language
that is preferred for the fields. Special value "all" (the default) can be
used to return a concatenated string of all available languages.
\code{\link[=gather_languages]{gather_languages()}} will select other available languages
in alphabetical order if the selected language is not available.}

\item{silent}{TRUE/FALSE: Should gather_languages() issue messages
and warnings?}
}
\value{
A character vector containing the values in the preferred language.
}
\description{
This function extracts the values for a preferred language
from a list containing values in multiple languages.
}
\examples{
\dontrun{
input_list <- list(list("en" = "English text", "de" = "Deutscher Text"),
                   list("en" = "Another english text", "de" = "Weiterer dt. Text"))
gather_languages(input_list, language = "de")
}
}
\seealso{
\itemize{
\item This function is used by: \code{\link[=simplify_idaifield]{simplify_idaifield()}}
}
}
