% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{get_configuration}
\alias{get_configuration}
\title{Get the Custom Project Configuration as Provided by the Field Desktop API}
\usage{
get_configuration(connection)
}
\arguments{
\item{connection}{A connection object as returned by \code{\link[=connect_idaifield]{connect_idaifield()}}}
}
\value{
A list of class \code{idaifield_config} containing the project
configuration, with \code{connection} and \code{projectname} stored as attributes.
Returns \code{NA} with a warning if the configuration could not be found or
the connection failed.
}
\description{
This function retrieves the complete project configuration (if existent)
from an \href{https://github.com/dainst/idai-field}{iDAI.field} project via
Field Desktop's configuration endpoint. The list will contain the complete
configuration as used in the project, including custom \emph{fields},
\emph{valuelists} and \emph{translations}.
}
\examples{
\dontrun{
conn <- connect_idaifield(
  serverip = "localhost", pwd = "hallo", project = "rtest"
)
config <- get_configuration(connection = conn)
}
}
\seealso{
\itemize{
\item Get the inputTypes from a Configuration: \code{\link[=parse_field_inputtypes]{parse_field_inputtypes()}}
\item This function is used by: \code{\link[=simplify_idaifield]{simplify_idaifield()}}.
\item \href{https://field.idai.world/manual}{iDAI.field Manual: 12. API}
}
}
