% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{get_field_inputtypes}
\alias{get_field_inputtypes}
\title{Produce a Matrix of Field \emph{inputTypes} from the Custom Project Configuration}
\usage{
get_field_inputtypes(config = NULL)
}
\arguments{
\item{config}{An \code{idaifield_config} as returned by \code{\link[=get_configuration]{get_configuration()}}}
}
\value{
A data.frame with four columns: category: each category, its parent
(or itself if it is a supercategory),
}
\description{
This function retrieves a matrix containing the \emph{inputTypes} of all
custom fields from a project configuration of an
\href{https://github.com/dainst/idai-field}{iDAI.field} project.
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo",
                          project = "rtest")
config <- get_configuration(connection = conn)
input_type_df <- get_field_inputtypes(config = config)
}
}
\seealso{
\itemize{
\item \code{\link[=get_configuration]{get_configuration()}}, \code{\link[=convert_to_onehot]{convert_to_onehot()}}
}
}
\keyword{internal}
