% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{get_idaifield_connection}
\alias{get_idaifield_connection}
\title{get_idaifield_connection}
\usage{
get_idaifield_connection(
  serverip = "192.168.1.199",
  user = "Anna Allgemeinperson",
  pwd = "password"
)
}
\arguments{
\item{projectname}{}
}
\value{
connection object
}
\description{
connects to the idaifield-database running on the serverip, is a shorter wrapper for sofa's Cushion$new
with defaults for idaifield. people still need to enter the serverip, arbitrary username and password.
}
