% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_language_lookup.R
\name{get_language_lookup}
\alias{get_language_lookup}
\title{Prepare a Language List as a Lookup Table}
\usage{
get_language_lookup(lang_list, language = "en", remove_config_names = TRUE)
}
\arguments{
\item{lang_list}{A list in the format used by iDAI.fields configuration,
containing a separate list for each language with its short
name (e.g. "en", "de") in which the "commons", "categories" etc. lists
are contained. Can be obtained with \code{\link[=get_configuration]{get_configuration()}}.}

\item{language}{Language short name that is to be extracted, e.g. "en",
defaults to "en"}
}
\value{
A data.frame that can serve as a lookup table, with the background
name in the "var" column, and the selected language in the "label" column.
}
\description{
The function compiles a table of background values and their translations
in the language selected from the configuration supplied to it.
}
\details{
Be aware: if two things have the same name
in the background of the database / project configuration but you use
different translations this will result in only one of the
translations being used.
}
\examples{
\dontrun{
conn <- connect_idaifield(serverip = "localhost",
                          project = "rtest",
                          pwd = "hallo")
config <- get_configuration(connection = conn)
lookup <- get_language_lookup(config$languages, language = "en")
}
}
\seealso{
\itemize{
\item Get the necessary configuration: \code{\link[=get_configuration]{get_configuration()}}
}
}
