% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{get_uid_type_list}
\alias{get_uid_type_list}
\title{get_uid_type_list}
\usage{
get_uid_type_list(idaifield_docs, verbose = FALSE)
}
\arguments{
\item{verbose}{TRUE or FALSE, TRUE returns a list including identifier and shorttitle which is more convenient for humans, and FALSE returns only UID and type, which is enough for internal selections}
}
\value{

}
\description{
gets a list of UIDs and Type for selecting later on
}
