% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importlist.R
\name{idaifield_as_df}
\alias{idaifield_as_df}
\title{idaifield_as_df}
\usage{
idaifield_as_df(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs(...)}
}
\value{

}
\description{
Converts the list that is returned by get_idaifield_docs(...) into a data.frame with
all columns present in the data. Individual rows will contain NA if there was no information
entered into the respective field in `idaifield`. Lists are currently not preserved (e.g.
the color-lists etc, but instead converted to a ;-seperated character string). They can be
pulled apart again, and I am also thinking about preserving them in this function (but have not
done that yet.)
}
\details{
If the list containing all meta-info (i.e. the docs-list)
is handed to the function it will automatically unnest to resource level).
}
