% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idaifieldr_demodata}
\alias{idaifieldr_demodata}
\title{idaifieldR Demodata}
\format{
## `idaifieldr_demodata`
A list (idaifield_docs) with 108 elements:
\describe{
  \item{identifier}{identifier / Kurzbezeichner}
  \item{all other values}{entries in the field database}
  ...
}
}
\source{
Random collection of imagined data
}
\usage{
idaifieldr_demodata
}
\description{
Test and Demonstration Data to illustrate the packages functionality, has
no real world representation.
}
\keyword{datasets}
