% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_changes.R
\name{idf_get_changes}
\alias{idf_get_changes}
\title{See when and by whom selected resources have been created or modified}
\usage{
idf_get_changes(connection, ids)
}
\arguments{
\item{connection}{A connection object as returned by \code{\link[=connect_idaifield]{connect_idaifield()}}}

\item{ids}{A vector of either \strong{identifier}s or \strong{UUID}s of resources for which
the changes should be returned.}
}
\value{
A data.frame in which each row corresponds to a change made to the
resource named in the \emph{identifier} column at the time stated in the \emph{date}
column by the user stated in the \emph{user} column. In the \emph{action} column,
the value can either be \emph{created} referring to the date of first creation
of the resource, or \emph{modified} referring to the date when it was changed.
}
\description{
See when and by whom selected resources have been created or modified
}
\examples{
\dontrun{
connection <- connect_idaifield(pwd = "hallo", project = "rtest")
index <- get_field_index(connection)
last_changed <- idf_last_changed(
    connection = connection,
    index = index,
    n = 100
)
changes_df <- idf_get_changes(connection, ids = last_changed)
}
}
\seealso{
\itemize{
\item \code{\link[=idf_last_changed]{idf_last_changed()}}
}
}
