% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_json_query}
\alias{idf_json_query}
\title{Query \emph{docs} from an iDAI.field Database Based on a JSON-Query}
\usage{
idf_json_query(connection, query, projectname = NULL)
}
\arguments{
\item{connection}{A connection object as returned
by \code{\link[=connect_idaifield]{connect_idaifield()}}}

\item{query}{A valid JSON-query as detailed in the relevant section of the
\href{https://docs.couchdb.org/en/stable/api/database/find.html}{CouchDB-API}
documentation.}

\item{projectname}{(deprecated) The name of the project to be queried (overrides
the one listed in the connection-object).}
}
\value{
An \code{idaifield_docs} list
}
\description{
This function can be used to gather \emph{docs} from an iDAI.field / Field
Desktop Database using a JSON-Query as detailed by the
\href{https://docs.couchdb.org/en/stable/api/database/find.html}{CouchDB-API Docs}.
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo", project = "rtest")

# Get all documents that contain "Anna Allgemeinperson" as processor
query <- '{ "selector": { "resource.processor": ["Anna Allgemeinperson"] } }'
result <- idf_json_query(conn, query = query)


# Get all documents where hasRestoration is TRUE
query <- '{ "selector": { "resource.hasRestoration": true } }'
result <- idf_json_query(conn, query = query)

# Get all documents where hasRestoration is TRUE
# but only the fields *identifier* and *hasRestoration*
query <- '{ "selector": { "resource.hasRestoration": true },
"fields": [ "resource.identifier", "resource.hasRestoration" ] }'
result <- idf_json_query(conn, query = query)

}
}
\seealso{
\itemize{
\item Learn how to build the selector-query with the
\href{https://docs.couchdb.org/en/stable/api/database/find.html}{CouchDB-API}.
\item Alternative functions: \code{\link[=idf_query]{idf_query()}}, \code{\link[=idf_index_query]{idf_index_query()}}
}
}
