% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_changes.R
\name{idf_last_changed}
\alias{idf_last_changed}
\title{Get a vector of UUIDs or identifiers of the last n changed resources}
\usage{
idf_last_changed(connection, index = NULL, n = 100)
}
\arguments{
\item{connection}{A connection object as returned
by \code{\link[=connect_idaifield]{connect_idaifield()}}}

\item{index}{A data.frame as returned by \code{\link[=get_field_index]{get_field_index()}}
(or \code{\link[=get_uid_list]{get_uid_list()}}).}

\item{n}{numeric. Maximum number of last changed resources to get.}
}
\value{
A vector of \code{UUID}s or \code{identifier}s of the most
recently changed \code{n} resources.
}
\description{
Retrieves the names/identifiers or UUIDs of the most recently changed
resources in the database. If an index as returned by \code{\link[=get_field_index]{get_field_index()}}
or \code{\link[=get_uid_list]{get_uid_list()}} is returned, the UUIDs are replaced by identifiers.
If not, the UUIDs are returned directly and can be used for querying e.g.
with
}
\examples{
\dontrun{
connection <- connect_idaifield(pwd = "hallo", project = "rtest")
index <- get_field_index(connection)
last_changed <- idf_last_changed(
    connection = connection,
    index = index,
    n = 100
)
}
}
