% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_idaifield.R
\name{idf_ping}
\alias{idf_ping}
\title{Ping the Field-Database}
\usage{
idf_ping(conn)
}
\arguments{
\item{conn}{The connection settings as returned by `connect_idaifield()` or
a direct Client as returned by `proj_idf_client()` for internal functions}
}
\value{
returns TRUE if connection works, FALSE if failed to connect
}
\description{
Helper for all functions that access the database to display
specific warning messages.
}
\examples{
\dontrun{
idaifield_connection <- connect_idaifield(ping = FALSE)
idf_ping(idaifield_connection)
}

}
