% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_projects.R
\name{idf_projects}
\alias{idf_projects}
\title{Get a Vector of Project Names Present in the Database}
\usage{
idf_projects(connection)
}
\arguments{
\item{connection}{The connection settings as
returned by \code{\link[=connect_idaifield]{connect_idaifield()}}}
}
\value{
A character vector containing the projects present in the database,
NA if connection was refused.
}
\description{
Get a Vector of Project Names Present in the Database
}
\examples{
\dontrun{
 connection <- connect_idaifield(pwd = "hallo")
 idf_project_list(connection)
}
}
\seealso{
\itemize{
\item \code{\link[=connect_idaifield]{connect_idaifield()}}, \code{\link[=idf_ping]{idf_ping()}}
}
}
