% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_query}
\alias{idf_query}
\title{idf_query}
\usage{
idf_query(connection, project, field = "type", value = "Brick")
}
\arguments{
\item{connection}{A connection object as returned by `connect_idaifield()`}

\item{project}{The name of the project to be queried.}

\item{field}{The resource field that should be selected for (i.e. "type" for
the type of resource (Pottery, Brick, Layer)).}

\item{value}{The value to be selected for in the specified field (i.e.
"Brick" when looking for resourced of type "Brick").}
}
\value{
A nested list with all matching docs and their metadata.
}
\description{
Queries resources from an iDAI.field 2-database that is currently running
}
\examples{
\dontrun{
idf_query(connection, project = "milet", field = "type", value = "Brick")
}

}
