% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{idf_sepdim}
\alias{idf_sepdim}
\title{Break down a list from a dimension field to a single value}
\usage{
idf_sepdim(dimensionList, name = "dimensionLength")
}
\arguments{
\item{dimensionList}{A list from one of the measurement fields
(dimensionLength, dimensionWidth, etc.) from a single resource (element).}

\item{name}{The name of the corresponding dimension List.}
}
\value{
A list containing simple values for each measured dimension from
the list; note: if a range was entered, it returns the mean without further
comment.
}
\description{
This function is a helper to `simplify_single_resource()`.
}
\examples{
\dontrun{
dimensionLength_new <- idf_sepdim(idaifield[[1]]$dimensionLength,
"dimensionLength")
}
}
\keyword{internal}
