% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_idaifield.R
\name{idf_show_categories}
\alias{idf_show_categories}
\title{Show the \emph{categories} Present in an \code{idaifield_docs}- or
\code{idaifield_resources}-list}
\usage{
idf_show_categories(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An an \code{idaifield_docs}- or \code{idaifield_resources}-list
as returned by \code{\link[=get_idaifield_docs]{get_idaifield_docs()}} or \code{\link[=idf_query]{idf_query()}}, \code{\link[=idf_index_query]{idf_index_query()}}
and \code{\link[=idf_json_query]{idf_json_query()}}.}
}
\value{
A character vector with the unique categories present in
the iDAI.field 2 / Field Desktop database the list was imported from.
}
\description{
Returns a list of all \emph{categories} present in the
iDAI.field 2 / Field Desktop database the list was imported from.
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
                                user = "R",
                                pwd = "hallo",
                                project = "rtest")

idaifield_docs <- get_idaifield_docs(connection = connection)

idf_show_categories(idaifield_docs)
}

}
