% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ifnotempty}
\alias{ifnotempty}
\title{Returns NA if an objects is empty}
\usage{
ifnotempty(item)
}
\arguments{
\item{item}{can be anything.}
}
\value{

}
\description{
(defense against empty list items from idaifield)
does basically nothing, just checking if input is of length 0 and if so
returning NA, if not returning the input untouched.
}
