% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_geometry.R
\name{maybe_to_json}
\alias{maybe_to_json}
\title{Convert a List to a JSON-String}
\usage{
maybe_to_json(x)
}
\arguments{
\item{x}{A list that should be converted (back) to a JSON string.}
}
\value{
A JSON-string
}
\description{
You can use \code{sf::st_read(geojson_string, quiet = TRUE)} to convert this to
a geometry you can plot in R.
}
\seealso{
\itemize{
\item This function is used in: \code{\link[=simplify_idaifield]{simplify_idaifield()}}
}
}
\keyword{internal}
