% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe_unnest_docs.R
\name{maybe_unnest_docs}
\alias{maybe_unnest_docs}
\title{(Maybe) Unnests an \code{idaifield_docs}-list}
\usage{
maybe_unnest_docs(x)
}
\arguments{
\item{x}{An object of class \code{idaifield_docs} to be processed. Will return
objects of other classes unchanged.}
}
\value{
If already unnested, the same object as handed to it. If not,
the same list with the top-level lists removed down to the "resource"-level,
as an \code{idaifield_resources}.
}
\description{
Checks if the input object is an \code{idaifield_docs}-list.
If so, the function unnests it by stripping all top-level lists and
returning only the list called "resource" within the db docs.
Any other class will simply be returned unchanged.
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo", project = "rtest")
idaifield_docs <- get_idaifield_docs(conn, raw = TRUE)
idaifield_resources <- maybe_unnest_docs(idaifield_docs)
}
}
