% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_if_empty}
\alias{na_if_empty}
\title{Returns NA if an object handed to the function is empty}
\usage{
na_if_empty(item)
}
\arguments{
\item{item}{any object whatsoever}
}
\value{
NA if empty, or the object that has been handed to it
}
\description{
This is a helper function in defence against empty list items from
iDAI.field 2 / Field Desktop, which sometimes occur.
It simply writes NA in the corresponding field if a list or any
kind of object handed to it is of length 0. Otherwise,
it returns the input untouched.
}
\examples{
\dontrun{
na_if_empty(1)
na_if_empty(list(2,3,4,list(4,5,4)))
na_if_empty(NULL)
}
}
\keyword{internal}
