% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{name_docs_list}
\alias{name_docs_list}
\title{Name the list of \code{idaifield_docs} by identifier}
\usage{
name_docs_list(docs)
}
\arguments{
\item{docs}{idaifield_docs}
}
\value{
same, but named
}
\description{
Name the list of \code{idaifield_docs} by identifier
}
\keyword{internal}
