% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reduce_relations}
\alias{reduce_relations}
\title{Reduces a (relations) list to its first element and warns accordingly.}
\usage{
reduce_relations(relation, uuid = NA, identifier = NA)
}
\arguments{
\item{uuid}{The id/UUID of the resource (for warning)}

\item{identifier}{The identifier of the resource (for warning)}

\item{resource}{The resource to be checked}
}
\value{
The reduced relations list
}
\description{
Reduces a (relations) list to its first element and warns accordingly.
}
\keyword{internal}
