% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_geometry.R
\name{reformat_geometry}
\alias{reformat_geometry}
\title{reformat the geometry of an iDAI.field 2 resource}
\usage{
reformat_geometry(geometry)
}
\arguments{
\item{geometry}{the list in docs$resource$geometry}
}
\value{
the same data in a more handy format that can be processed with
tools like the sp-package to produce polygons etc.
}
\description{
reformat the geometry of an iDAI.field 2 resource
}
