% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{remove_config_names}
\alias{remove_config_names}
\title{Remove Everything Before the Colon in a Character Vector}
\usage{
remove_config_names(
  conf_names = c("identifier", "configname:test", "test"),
  silent = FALSE
)
}
\arguments{
\item{conf_names}{A character vector.}

\item{silent}{Should the message that duplicates were
detected be suppressed? Default is FALSE.}
}
\value{
The same character vector with everything before
the first colon (including the colon) removed.
}
\description{
Removes everything before the first colon (including the colon)
in a character vector. Used for cleaning up category or field names that
have been produced using the \emph{Project Configuration Editor} in Field Desktop.
}
\examples{
\dontrun{
nameslist <- c(
  "relation.liesWithin",
  "relation.liesWithinLayer",
  "campaign.2022",
  "rtest:test",
  "pergamon:neuesFeld",
  "neuesFeld"
)
remove_config_names(nameslist, silent = FALSE)
}
}
