% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reorder_colnames}
\alias{reorder_colnames}
\title{Reorder the column names for `idaifield_as_matrix()`}
\usage{
reorder_colnames(colnames, order = "default")
}
\arguments{
\item{colnames}{a character vector with colnames}

\item{order}{either "default" for default order (first columns are
"identifier", "type", "shortDescription" and the rest is as assembled) or
a character vector with exact column names that will then place these as
the first n columns of the matrix produced by idaifield_as_matrix()}
}
\value{
a character vector
}
\description{
Reorder the column names for `idaifield_as_matrix()`
}
\examples{
\dontrun{
colnames <- c("materialType", "identifier", "shortDescription", "type")

reorder_colnames(colnames, order = "default")
}
}
\keyword{internal}
