% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{response_to_list}
\alias{response_to_list}
\title{Format the response of a crul::HttpClient from json to list}
\usage{
response_to_list(response = NULL)
}
\arguments{
\item{response}{"HttpResponse" as returned by client$get where
client was produced by `proj_idf_client()`}
}
\value{
a list
}
\description{
Format the response of a crul::HttpClient from json to list
}
\keyword{internal}
