% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_idaifield.R
\name{show_categories}
\alias{show_categories}
\title{show_categories}
\usage{
show_categories(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object as returned by `get_idaifield_docs(...)`;
unnests to resource level if it didn't already happen.}
}
\value{
a character vector with the types represented in
the idaifield_docs/resource
}
\description{
Returns a list of types present in the iDAI.field 2 / Field Desktop database.
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
                                user = "R",
                                pwd = "hallo",
                                project = "rtest")
idaifield_docs <- get_idaifield_docs(connection = connection)

show_categories(idaifield_docs)
}
}
