% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importlist.R
\name{show_type_list}
\alias{show_type_list}
\title{show_type_list}
\usage{
show_type_list(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs(...); unnests to
resource level if it didn't already happen.}
}
\value{

}
\description{
Returns a list of types present in the idaifield-database (for orientation)
it would be nice to be able to embed the translations, but i am guessing thats only
possible if the configuration was available
}
