% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_idaifield}
\alias{simplify_idaifield}
\title{Simplify a List Imported from an iDAI.field / Field Desktop-\code{1041-1}#' Simplify a List Imported from an iDAI.field / Field Desktop-Database}
\usage{
simplify_idaifield(
  resources,
  index = NULL,
  config = NULL,
  silent = FALSE,
  language = "all",
  keep_geometry = FALSE
)
}
\arguments{
\item{index}{If NULL (default) the list of UUIDs and identifiers is
automatically generated within this function using \code{\link[=get_uid_list]{get_uid_list()}}. This only makes sense if
the list handed to \code{\link[=simplify_idaifield]{simplify_idaifield()}} had not been selected yet. If it
has been, you should supply a data.frame as returned
by \code{\link[=get_field_index]{get_field_index()}}.}

\item{config}{A configuration list as returned
by \code{\link[=get_configuration]{get_configuration()}}}

\item{silent}{TRUE/FALSE, default: FALSE. Should messages be suppressed?}

\item{language}{The short name (e.g. "en", "de", "fr") of the language
that is preferred for the fields. Special value "all" (the default) can be
used to return a concatenated string of all available languages.
\code{\link[=gather_languages]{gather_languages()}} will select other available languages
in alphabetical order if the selected language is not available.}

\item{idaifield_docs}{An \code{idaifield_docs} or \code{idaifield_resources}-list as
returned by \code{\link[=get_idaifield_docs]{get_idaifield_docs()}} or \code{\link[=idf_query]{idf_query()}},
\code{\link[=idf_index_query]{idf_index_query()}}, and \code{\link[=idf_json_query]{idf_json_query()}}.}
}
\value{
An \code{idaifield_simple}-list containing the same resources in
a different format depending on the parameters used.
}
\description{
The function will take a list as returned by
\code{\link[=get_idaifield_docs]{get_idaifield_docs()}}, \code{\link[=idf_query]{idf_query()}}, \code{\link[=idf_index_query]{idf_index_query()}}, or
\code{\link[=idf_json_query]{idf_json_query()}} and process it to make the list more usable.
It will unnest a few lists, including the dimension-lists and the
period-list to provide single values for later processing with
\code{\link[=idaifield_as_matrix]{idaifield_as_matrix()}}.
If a connection to the database can be established, the function will
get the relevant project configuration and convert custom checkboxes-fields
to multiple lists, each for every value from the respective valuelist,
to make them more accessible during the conversion with
\code{\link[=idaifield_as_matrix]{idaifield_as_matrix()}}.
It will also remove the custom configuration field names that are in use
since iDAI.field 3 / Field Desktop and consist of "projectname:fieldName".
Only the "projectname:"-part will be removed.
}
\details{
Please note: The function will need an Index (i.e. uidlist as provided
by \code{\link[=get_uid_list]{get_uid_list()}}) of the complete project database to correctly replace
the UUIDs with their corresponding identifiers! Especially if a selected
list is passed to \code{\link[=simplify_idaifield]{simplify_idaifield()}}, you need to supply the uidlist
of the complete project database as well.

Formatting of various lists: Dimension measurements as well as dating are
reformatted and might produce unexpected results.
For the dating, all begin and end values are evaluated and for each resource,
the minimum value from "begin" and maximum value from "end" is selected.
For the dimension-fields, if a ranged measurement was selected, a mean
will be returned.
}
\examples{
\dontrun{
connection <- connect_idaifield(
    serverip = "localhost",
    project = "rtest",
    pwd = "hallo"
)
idaifield_docs <- get_idaifield_docs(connection = connection)

simpler_idaifield <- simplify_idaifield(idaifield_docs)
}
}
\seealso{
\itemize{
\item This function uses: \code{\link[=idf_sepdim]{idf_sepdim()}}, \code{\link[=remove_config_names]{remove_config_names()}}
\item When find_layers = TRUE: \code{\link[=find_layer]{find_layer()}}, this only works when the function can get an index/uidlist!
\item Depending on the \code{replace_uids}-argument: \code{\link[=fix_relations]{fix_relations()}} with \code{\link[=replace_uid]{replace_uid()}}
\item If \code{uidlist = NULL}: \code{\link[=get_uid_list]{get_uid_list()}}
}
}
