% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_single_resource}
\alias{simplify_single_resource}
\title{Simplifies a single resource from the iDAI.field 2 / Field Desktop Database}
\usage{
simplify_single_resource(
  resource,
  index = NULL,
  config = NULL,
  replace_uids = TRUE,
  find_layers = TRUE,
  keep_geometry = TRUE,
  language = "all",
  silent = FALSE
)
}
\arguments{
\item{resource}{One resource (element) from an \code{idaifield_resources}-list.}

\item{index}{If NULL (default) the list of UUIDs and identifiers is
automatically generated within this function using \code{\link[=get_uid_list]{get_uid_list()}}. This only makes sense if
the list handed to \code{\link[=simplify_idaifield]{simplify_idaifield()}} had not been selected yet. If it
has been, you should supply a data.frame as returned
by \code{\link[=get_field_index]{get_field_index()}}.}

\item{config}{A configuration list as returned
by \code{\link[=get_configuration]{get_configuration()}}}

\item{replace_uids}{TRUE/FALSE: Should UUIDs be automatically replaced with the
corresponding identifiers? Defaults is TRUE. Uses: \code{\link[=fix_relations]{fix_relations()}} with
\code{\link[=replace_uid]{replace_uid()}}, and also: \code{\link[=find_layer]{find_layer()}}}

\item{find_layers}{TRUE/FALSE. Default is FALSE. If TRUE, adds another column
with the 'Layer' (see \code{getOption("idaifield_categories")$layers}, can be
modified) in which a resource is contained  recursively. That means that
even if it does not immediately lie within this layer, but is
contained by one or several other resources in said layer, a new column
("liesWithinLayer") will still show the layer.
Example: A sample "A" in Find "001" from layer "Layer1" will
usually have "001" as the value in "liesWithin". With find_layers, there will
be another column called "liesWithinLayer" which contains "Layer1" for both
sample "A" and Find "001".}

\item{keep_geometry}{Should the geometry of each resource be kept?}

\item{language}{The short name (e.g. "en", "de", "fr") of the language
that is preferred for the fields. Special value "all" (the default) can be
used to return a concatenated string of all available languages.
\code{\link[=gather_languages]{gather_languages()}} will select other available languages
in alphabetical order if the selected language is not available.}

\item{silent}{TRUE/FALSE, default: FALSE. Should messages be suppressed?}
}
\value{
A single resource (element) for an \code{idaifield_resources}-list.
}
\description{
This function is a helper to \code{simplify_idaifield()}.
}
\examples{
\dontrun{
simpler_resource <- simplify_single_resource(resource,
replace_uids = TRUE,
uidlist = uidlist,
keep_geometry = FALSE)
}
}
\keyword{internal}
