% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{stop_if_not_idf_connection_settings}
\alias{stop_if_not_idf_connection_settings}
\title{Check if object is a "idf_connection_settings" and stop if not.}
\usage{
stop_if_not_idf_connection_settings(x)
}
\arguments{
\item{x}{an object}
}
\value{
nothing
}
\description{
Check if object is a "idf_connection_settings" and stop if not.
}
\keyword{internal}
