% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{type_to_category}
\alias{type_to_category}
\title{Replace type with category in resource list names}
\usage{
type_to_category(docs)
}
\arguments{
\item{idaifield_docs}{}
}
\value{
the docs with "type" renamed to "category"
}
\description{
Replace type with category in resource list names
}
\keyword{internal}
