% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_docs.R
\name{unnest_docs}
\alias{unnest_docs}
\title{Unnesting a idaifield_docs-List down to resource level}
\usage{
unnest_docs(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{A list as provided by `get_idaifield_docs()`.
`get_idaifield_docs()` employs this function already when
setting `raw = FALSE`.}
}
\value{
a list of class idaifield_resources (same as idaifield_docs,
but the top-level with meta-information has been removed to make the actual
resource data more accessible)
}
\description{
This function unnests the lists provided by iDAI.field 2 / Field Desktop.
The actual data of a resource is usually stored in a sub-list
behind $doc$resource, which contains the data one would mostly
want to work with in R. The top level data contains information
about who created and modified the resource at what time and
is irrelevant for any analysis of the database contents itself.
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
idaifield_docs <- get_idaifield_docs(connection = connection,
projectname = "rtest", simplified = FALSE)

idaifield_resources <- unnest_docs(idaifield_docs)
}
}
\keyword{internal}
