% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{unnest_resource}
\alias{unnest_resource}
\title{removes the top layers of the json-list since they are not needed}
\usage{
unnest_resource(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{}
}
\value{

}
\description{
removes the top layers of the json-list since they are not needed
}
