% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{warn_for_project}
\alias{warn_for_project}
\title{TEMPORARY FUNCTION to warn about project(name)}
\usage{
warn_for_project(project = NULL, fail = FALSE)
}
\arguments{
\item{project}{connection$project}

\item{fail}{should there be an error?}
}
\value{
TRUE/FALSE
}
\description{
TEMPORARY FUNCTION to warn about project(name)
}
\keyword{internal}
