% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{bce_ce}
\alias{bce_ce}
\title{Translate a list for one dating value from iDAI.field to a positive or negative number}
\usage{
bce_ce(list)
}
\arguments{
\item{list}{A named list containing (at least): inputYear (number),
and inputType ("bce", "ce", "bp")}
}
\value{
The year as a number, negative when BCE, positive when CE
}
\description{
(Field does save numbers in this format, but apparently not all of them.
This corrects for wrong numbers. Numbers bp/before present are subtracted
from 1950 to get dates BCE.)
}
\examples{
\dontrun{
list <- list(inputYear = 100, inputType = "bce")
bce_ce(list)
}

}
\keyword{internal}
