% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_and_unnest}
\alias{check_and_unnest}
\title{Check and unnest a list}
\usage{
check_and_unnest(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object to be used by one of the
functions in this package}
}
\value{
if already unnested, the same object as handed to it. if not,
the same list with the toplevel removed.
}
\description{
Checks for list of class "idaifield_docs" and if the object is already
unnested (i.e. of class "idaifield_resources"); if it is not, does so.
If it cannot be processed, because it is not an idaifield_docs or
idaifield_resources object, throws error.
TODO: wrap in trycatch to process anyway if possible
}
\examples{
\dontrun{
idaifield_docs <- get_idaifield_docs(projectname = "rtest",
connection = connect_idaifield(serverip = "192.168.1.21",
user = "R",
pwd = "password"))

check_and_unnest(idaifield_docs)
}
}
