% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_and_unnest}
\alias{check_and_unnest}
\title{check_and_unnest}
\usage{
check_and_unnest(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object to be used by one of the
functions in this package}
}
\value{
if already unnested, the same object as handed to it. if not,
the same list with the toplevel removed.
}
\description{
Checks if the object is already unnested, and if it isnt, does so.
If it cant be processed, because it is not an idaifield_docs or
idaifield_resources object, throws error.
}
\examples{
\dontrun{
idaifield_docs <- get_idaifield_docs(serverip = "192.168.1.21",
projectname = "testproj",
user = "R",
pwd = "password")

check_and_unnest(idaifield_docs)
}
}
