% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_for_sublist}
\alias{check_for_sublist}
\title{check_for_sublist}
\usage{
check_for_sublist(single_resource_field)
}
\arguments{
\item{single_resource_field}{a list to be checked}
}
\value{
TRUE/FALSE
}
\description{
Checks if a list has sublists and returns TRUE if so
}
