% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_if_uid}
\alias{check_if_uid}
\title{check_if_uid}
\usage{
check_if_uid(string)
}
\arguments{
\item{string}{A character string or vector of character strings that should
be checked for being a UID as used in iDAI.field 2 / Field Desktop}
}
\value{
a vector of the same length as string containing TRUE if
the corresponding item in string is a UID, FALSE if not
}
\description{
check_if_uid
}
\examples{
\dontrun{
check_if_uid(string = "0324141a-8201-c5dc-631b-4dded4552ac4")
check_if_uid(string = "not a uid")
}
}
\keyword{internal}
