% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_idaifield.R
\name{connect_idaifield}
\alias{connect_idaifield}
\title{Establishes a connection to the iDAI.field 2 Client.}
\usage{
connect_idaifield(
  serverip = "127.0.0.1",
  user = "Anna Allgemeinperson",
  pwd = "password"
)
}
\arguments{
\item{serverip}{The IP that the user can find in iDAI.field 2's settings as
"Eigene Adresse" without the port-specification}

\item{user}{A user name (anything works, really.)}

\item{pwd}{The Password as it is displayed in the iDAI.field 2-settings
as "Eigenes Passwort"}
}
\value{
a connection object
}
\description{
Establishes a connection to the iDAI.field 2 Client.
}
\examples{
conn <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
}
