% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{convert_to_onehot}
\alias{convert_to_onehot}
\title{Break down a list from a checkbox field to onehot-coded values}
\usage{
convert_to_onehot(resource, fieldtypes)
}
\arguments{
\item{resource}{A list from one of the fields that can have multiple values
from a single resource (element).}

\item{fieldtypes}{a matrix of fields with the given inputType as
returned by `get_field_inputtypes()`}
}
\value{
The resource object with the values of checkboxes
separated into one-hot-coded versions.
}
\description{
This function is a helper function to
`simplify_idaifield()` that takes a list from one of the
fields marked in the config as containing checkboxes and converts
the list to onehot-coded values.
}
\examples{
\dontrun{
...
}
}
\seealso{
\code{\link{simplify_idaifield}}, \code{\link{get_field_inputtypes}}
}
