% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{convert_to_onehot}
\alias{convert_to_onehot}
\title{Break down a list from a checkbox field to onehot-coded values}
\usage{
convert_to_onehot(resource, fieldtypes)
}
\arguments{
\item{resource}{A list from one of the measurement fields
(dimensionLength, dimensionWidth, etc.) from a single resource (element).}

\item{config}{A configuration list as returned by `get_configuration()`}
}
\value{
The resource object with the checkboxes seperated
}
\description{
This function is a helper to `simplify_single_resource()`.
}
\examples{
\dontrun{
conn <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
config <- get_configuration(connection = conn,
projectname = "rtest")
idaifield_resources[[1]] <- convert_to_onehot(config,
idaifield_resources[[1]])
}
}
\keyword{internal}
