% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Loads the iDAI.field Categories}
\usage{
.onLoad(lib, pkg)
}
\arguments{
\item{libname}{The name of the library.}

\item{pkgname}{The name of the package.}
}
\value{
A list of options.
}
\description{
This function loads the iDAI.field categories with the default values.
}
