% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{extract_inputtypes}
\alias{extract_inputtypes}
\title{Title}
\usage{
extract_inputtypes(nested_list, parent_name = NULL, category_name = NULL)
}
\arguments{
\item{category_name}{}
}
\description{
Title
}
