% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_layer.R
\name{find_layer}
\alias{find_layer}
\title{Find the Layer a resource is contained in}
\usage{
find_layer(
  resource = resource,
  uidlist = NULL,
  liesWithin = NULL,
  strict = FALSE,
  opts = NULL
)
}
\arguments{
\item{resource}{One resource (element) from an idaifield_resources-list.}

\item{uidlist}{A data.frame as returned by `get_field_index()`.}

\item{liesWithin}{Only for recursion: a dataframe this function hands
to itself, otherwith usually NULL}

\item{strict}{TRUE/FALSE (currently only for testing)}
}
\value{
chr
}
\description{
Warning: recursive and currently no error handling
}
\details{
Helper to simplify_single_resource(). Traces the liesWithin fields to
find the one that is a Layer and returns the corresponding identifier.

#TODO: Somehow this is super convoluted.
}
\keyword{internal}
