% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_named_list}
\alias{find_named_list}
\title{Return a sub-list from a nested list by name}
\usage{
find_named_list(target_name, nested_list)
}
\arguments{
\item{target_name}{a character value expected to be the name of one list}

\item{nested_list}{a nested list}
}
\value{
The first sub-list that has the name passed to target_name.
}
\description{
Return a sub-list from a nested list by name
}
\examples{
list <- list(first1 = list(second1 = list(1, 2, 3), second2 = list(4, 5, 6)), first2 = list(7, 8, 9))
find_named_list("second2", list)
}
