% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_docs.R
\name{find_resource}
\alias{find_resource}
\title{Finding the "resource"-list in a CouchDB-output}
\usage{
find_resource(list)
}
\arguments{
\item{list}{a list formatted from a CouchDB-JSON-output}
}
\value{
the list of resource-lists
}
\description{
Finding the "resource"-list in a CouchDB-output
}
\keyword{"test",}
\keyword{"test2"))),}
\keyword{"test2")),}
\keyword{"warning")}
\keyword{<-}
\keyword{=}
\keyword{\dontrun{}
\keyword{find_resource(testlist)}
\keyword{internal}
\keyword{list(docs}
\keyword{list(list(resource}
\keyword{list(resource}
\keyword{list(test}
\keyword{test2}
\keyword{testlist}
\keyword{warning}
\keyword{}}
