% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{fix_dating}
\alias{fix_dating}
\title{Reduce the Dating-list to min/max-values}
\usage{
fix_dating(dat_list)
}
\arguments{
\item{dat_list}{The "dating"-list of any resource from
an `idaifield_...`-list}
}
\value{
a reformatted list, containing min and max dating and additional
info as well as the original values in the "comment"-element
}
\description{
Note: This function will evaluate all begin and end values for the
dating of the resource object and evaluate only their min and max values!
}
\examples{
\dontrun{
dat_list <- list(type = "range",
                 begin = list(inputYear = 2000, inputType = "bce"),
                 end = list(inputYear = 2000, inputType = "ce"))
fix_dating(dat_list)
}
}
\keyword{internal}
