% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_relations.R
\name{fix_relations}
\alias{fix_relations}
\title{Unnest the relations in a resource and replace the UUIDs with identifiers}
\usage{
fix_relations(resource, replace_uids = TRUE, uidlist = NULL)
}
\arguments{
\item{resource}{One item from a list of class idaifield_resources}

\item{replace_uids}{TRUE/FALSE. If TRUE, replaces the UUIDs in each relation
with the corresponding identifiers. If FALSE, just unnests the list.
Default is TRUE.}

\item{uidlist}{Only needs to be provided if replace_uids = TRUE.
A data.frame as returned by \code{\link{get_uid_list}} or
\code{\link{get_field_index}}.}
}
\value{
The same resource with its relations unnested (and replaced with
identifiers if `replace_uids` is set to `TRUE`).
}
\description{
The function will flatten the relations list to more non-nested lists with
with "relation."-prefix and replace the UUIDs in the lists values with
the corresponding identifiers from the uidlist using
\code{\link{replace_uid}} to make the result more readable.
}
\examples{
\dontrun{
new_relations_list <- fix_relations(resource,
replace_uids = FALSE,
uidlist = NULL)
}
}
