% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{gather_languages}
\alias{gather_languages}
\title{Gather multilanguage fields}
\usage{
gather_languages(input_list, language = "en", silent = FALSE)
}
\arguments{
\item{input_list}{a list with character values containing (or not)
sublists for each language}

\item{language}{the short name (e.g. "en", "de", "fr") of the language that
is preferred for the fields, defaults to english ("en")}

\item{silant}{TRUE/FALSE: Should gather_languages()
issue messages and warnings?}
}
\value{
a vector containing the values
}
\description{
Gather multilanguage fields
}
\examples{
\dontrun{
input_list <- list(list("en" = "English text", "de" = "Deutscher Text"),
                   list("en" = "Another english text", "de" = "Weiterer dt. Text"))
gather_languages(input_list, language = "de")
}
}
\keyword{internal}
