% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uid_list.R
\name{gather_trenches}
\alias{gather_trenches}
\title{Get a vector of \emph{Place}-resources each element from the index is located in}
\usage{
gather_trenches(uidlist)
}
\arguments{
\item{uidlist}{as returned by \code{\link[=get_uid_list]{get_uid_list()}}
and \code{\link[=get_field_index]{get_field_index()}}}
}
\value{
a vector containing the Place each resource is located in
}
\description{
Get a vector of \emph{Place}-resources each element from the index is located in
}
\examples{
\dontrun{
uidlist$Place <- gather_trenches(uidlist)
}
}
\keyword{internal}
