% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uid_list.R
\name{gather_trenches}
\alias{gather_trenches}
\title{Get a vector of places each element from the uidlist is located in}
\usage{
gather_trenches(uidlist)
}
\arguments{
\item{uidlist}{as returned by `get_uid_list()` and `get_field_index()`}
}
\value{
a vector of Place-resources
}
\description{
Get a vector of places each element from the uidlist is located in
}
\examples{
\dontrun{
uidlist$Place <- gather_trenches(uidlist)
}
}
\keyword{internal}
