% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{get_configuration}
\alias{get_configuration}
\title{Get the Custom Project Configuration as Provided by the Field API}
\usage{
get_configuration(connection)
}
\arguments{
\item{connection}{A connection object as returned by \code{\link[=connect_idaifield]{connect_idaifield()}}}
}
\value{
A list containing the project configuration; \code{NA} if the configuration
could not be found or the connection failed.
}
\description{
This function retrieves the complete project configuration (if existent)
from an \href{https://github.com/dainst/idai-field}{iDAI.field} project via
Field's configuration endpoint. The list will only contain the complete
configuration as used in the project, including custom and \emph{fields},
\emph{valuelists} and \emph{translations}.
}
\examples{
\dontrun{
conn <- connect_idaifield(
  serverip = "localhost", pwd = "hallo", project = "rtest"
)
config <- get_configuration(connection = conn)
}
}
\seealso{
\itemize{
\item Get the inputTypes from a Configuration: \code{\link[=extract_inputtypes]{extract_inputtypes()}}
\item This function is used by: \code{\link[=simplify_idaifield]{simplify_idaifield()}}.
}
}
