% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{get_field_inputtypes}
\alias{get_field_inputtypes}
\title{get_field_inputtypes: returns a matrix of inputTypes}
\usage{
get_field_inputtypes(config, inputType = "all")
}
\arguments{
\item{config}{A configuration list as returned by `get_configuration()`}

\item{inputType}{if specified, matrix is filtered to return only the
specified type}
}
\value{
a matrix of fields with the given inputType
}
\description{
This function retrieves a matrix containing the inputTypes of all
fields and their corresponding inputTypes
from a project configuration (if existent) of an
iDAI.field project.
}
\examples{
\dontrun{
conn <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
config <- get_configuration(connection = conn,
projectname = "rtest")
checkboxes <- get_field_inputtypes(config, inputType = "checkboxes")
}
}
