% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_configuration.R
\name{get_field_inputtypes}
\alias{get_field_inputtypes}
\title{Produce a Matrix of Field \emph{inputTypes} from the Custom Project Configuration}
\usage{
get_field_inputtypes(
  config,
  inputType = "all",
  remove_config_names = TRUE,
  silent = FALSE
)
}
\arguments{
\item{config}{A configuration list as returned
by \code{\link[=get_configuration]{get_configuration()}}}

\item{inputType}{If specified, matrix is filtered to return only the
specified type.}

\item{remove_config_names}{TRUE/FALSE: Should the name of the project be
removed from field names of the configuration? (Default is TRUE.)
(Should e.g.: \emph{test:amount} be renamed to \emph{amount},
see \code{\link[=remove_config_names]{remove_config_names()}}.)}

\item{silent}{TRUE/FALSE, default: FALSE. Should messages be suppressed?}
}
\value{
A matrix of fields (with the given \emph{inputType}).
}
\description{
This function retrieves a matrix containing the \emph{inputTypes} of all
custom fields from a project configuration of an
\href{https://github.com/dainst/idai-field}{iDAI.field} project.
}
\examples{
\dontrun{
conn <- connect_idaifield(serverip = "127.0.0.1",
                          pwd = "hallo",
                          project = "rtest")
config <- get_configuration(connection = conn)
checkboxes <- get_field_inputtypes(config, inputType = "checkboxes")
}
}
\seealso{
\itemize{
\item \code{\link[=get_configuration]{get_configuration()}}, \code{\link[=convert_to_onehot]{convert_to_onehot()}}
\item This function is used by: \code{\link[=simplify_idaifield]{simplify_idaifield()}}.
}
}
