% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{get_idaifield_docs}
\alias{get_idaifield_docs}
\title{get_idaifield_docs}
\usage{
get_idaifield_docs(
  idaifield_connection = get_idaifield_connection(),
  projectname = "projektname",
  all = TRUE
)
}
\arguments{
\item{all}{}
}
\value{

}
\description{
Imports the docs from an idaifield-database that is currently running and synching, needs a connection-object
as produced by sofa directly or with get_idaifield_connection()
}
