% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fundamentals.R
\name{get_idaifield_docs}
\alias{get_idaifield_docs}
\title{get_idaifield_docs}
\usage{
get_idaifield_docs(
  serverip = "192.168.1.199",
  user = "Anna Allgemeinperson",
  pwd = "password",
  projectname = "projektname",
  all = TRUE
)
}
\arguments{
\item{pwd}{}
}
\value{

}
\description{
Imports the docs from an idaifield-database that is currently running and synching,
needs params to be processed to idaifield_connection()
}
