% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_language_lookup.R
\name{get_language_lookup}
\alias{get_language_lookup}
\title{get_language_lookup(): Prepare a Language List as a Lookup Table}
\usage{
get_language_lookup(lang_list, language = "en")
}
\arguments{
\item{lang_list}{A list in the format used by iDAI.fields configuration,
containing a separate list for each language with its short
name (e.g. "en", "de") in which the "commons", "categories" etc. lists
are contained.}

\item{language}{Language short name that is to be extracted, e.g. "en",
defaults to "en"}
}
\value{
A data.frame that can serve as a lookup table, with the background
name in the "var" column, and the selected language in the "label" column.
}
\description{
get_language_lookup(): Prepare a Language List as a Lookup Table
}
\examples{
\dontrun{
conn <- connect_idaifield(serverip = "127.0.0.1",
                          user = "R", pwd = "hallo")
config <- get_configuration(connection = conn, projectname = "rtest")
lookup <- get_language_lookup(config$languages, language = "en")
}
}
