% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uid_list.R
\name{get_uid_list}
\alias{get_uid_list}
\title{get_uid_list}
\usage{
get_uid_list(idaifield_docs, verbose = FALSE, gather_trenches = FALSE)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs}

\item{verbose}{TRUE or FALSE (= anything but TRUE), TRUE returns a list
including identifier and shorttitle which is more convenient for humans,
and FALSE returns only UID and type, which is enough for internal selections}

\item{gather_trenches}{defaults to FALSE. If TRUE, adds another column that
records the Place each corresponding Trench and its sub-resources lie within.
(Useful for grouping the finds of several trenches.)}
}
\value{
a list of UIDs and their Types, Identifiers and shortDescriptions
}
\description{
All resources from `idaifield` are referenced with their UID in the relations
fields. Therefore, for many purposes a lookup-table needs to be provided
in order to get to the actual identifiers and names of the resources
referenced.
}
\details{
This function is also good for a quick overview / a list of all the
resources that exist along with their identifiers and short descriptions
and can be used to select the resources along their respective Types
(e.g. Pottery, Layer etc.). Please note that in any case the internal
names of everything will be used. If you relabeled `Trench` to `Schnitt` in
your language-configuration, this will still be `Trench` here.
None of these functions have any respect for language configuration!
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
idaifield_docs <- get_idaifield_docs(connection = connection,
projectname = "rtest", simplified = FALSE)

uidlist <- get_uid_list(idaifield_docs, verbose = TRUE)
}
}
