% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idaifield_as_df.R
\name{idaifield_as_df}
\alias{idaifield_as_df}
\title{idaifield_as_df}
\usage{
idaifield_as_df(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs(...)}
}
\value{
a data.frame
}
\description{
Converts the list that is returned by get_idaifield_docs(...) into a
data.frame with all columns present in the data. Individual rows will
contain NA if there was no information entered into the respective field
in iDAI.field 2. Lists are currently not preserved (e.g. the color-lists etc,
but instead converted to a ;-seperated character string). They can be pulled
apart again, and I am also thinking about preserving them in this
function (but have not done that yet.)
}
\details{
If the list containing all meta-info (i.e. the docs-list)
is handed to the function it will automatically unnest to resource level).
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "192.168.2.21",
user = "R", pwd = "hallo")
idaifield_docs <- get_idaifield_docs(connection,
projectname = "rtest", simplified = FALSE)

idaifield_df <- idaifield_as_df(idaifield_docs)
}
}
