% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_index_query}
\alias{idf_index_query}
\title{Query resources from an iDAI.field database based on the uidlist}
\usage{
idf_index_query(
  connection,
  field = "category",
  value = "Brick",
  uidlist = NULL,
  projectname = NULL
)
}
\arguments{
\item{connection}{A connection object as returned by `connect_idaifield()`}

\item{field}{character. The resource field that should be selected for
(options are limited to the columns names of the uidlist).}

\item{value}{character. The value to be selected for in the specified field.}

\item{uidlist}{A data.frame as returned by `get_field_index()`
(or `get_uid_list()`)}

\item{projectname}{The name of the project to be queried (overrides
the one listed in the connection-object).}
}
\value{
An 'idaifield_docs' list
}
\description{
Query resources from an iDAI.field database based on the uidlist
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo", project = "rtest")
uidlist <- get_field_index(conn)
idf_index_query(conn,
                field = "category",
                value = "Brick",
                uidlist = uidlist)
}

}
