% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_index_query}
\alias{idf_index_query}
\title{Query \emph{docs} from an iDAI.field Database Based on an Index (uidlist)}
\usage{
idf_index_query(
  connection,
  field = "category",
  value = "Brick",
  uidlist = NULL,
  projectname = NULL
)
}
\arguments{
\item{connection}{A connection object as returned
by \code{\link[=connect_idaifield]{connect_idaifield()}}}

\item{field}{character. The resource field that should be selected for
(options are limited to the column names of the uidlist).}

\item{value}{character. The value to be selected for in the specified field.}

\item{uidlist}{A data.frame as returned by \code{\link[=get_field_index]{get_field_index()}}
(or \code{\link[=get_uid_list]{get_uid_list()}}).}

\item{projectname}{(deprecated) The name of the project to be queried (overrides
the one listed in the connection-object).}
}
\value{
An \code{idaifield_docs} list
}
\description{
This function can be used to gather \emph{docs} from an iDAI.field / Field
Desktop Database according to the values of listed in an index as returned
by \code{\link[=get_field_index]{get_field_index()}} or \code{\link[=get_uid_list]{get_uid_list()}}.
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo", project = "rtest")
uidlist <- get_field_index(conn)
idf_index_query(conn,
                field = "category",
                value = "Brick",
                uidlist = uidlist)
}

}
\seealso{
\itemize{
\item Alternative functions: \code{\link[=idf_query]{idf_query()}}, \code{\link[=idf_json_query]{idf_json_query()}}
}
}
