% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_index_query}
\alias{idf_index_query}
\title{idf_index_query}
\usage{
idf_index_query(
  connection,
  projectname = "NULL",
  field = "type",
  value = "Brick",
  uidlist = NULL
)
}
\arguments{
\item{connection}{A connection object as returned by `connect_idaifield()`}

\item{projectname}{The name of the project to be queried.}

\item{field}{The resource field that should be selected for (options are
limited to the columns names of the uidlist).}

\item{value}{The value to be selected for in the specified field.}

\item{uidlist}{A data.frame as returned by `get_uid_list()`.}
}
\value{
An 'idaifield_docs' list
}
\description{
Queries resources from an iDAI.field database that is currently running
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo")
uidlist <- get_uid_list(conn, projectname = "rtest")
idf_index_query(conn,
projectname = "rtest",
field = "type",
value = "Brick",
uidlist = uidlist)
}

}
