% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_index_query}
\alias{idf_index_query}
\title{idf_index_query}
\usage{
idf_index_query(
  connection,
  project,
  field = "type",
  value = "Brick",
  keep_geometry = TRUE,
  uidlist = NULL
)
}
\arguments{
\item{connection}{A connection object as returned by `connect_idaifield()`}

\item{project}{The name of the project to be queried.}

\item{field}{The resource field that should be selected for (options are
limited to the columns names of the uidlist).}

\item{value}{The value to be selected for in the specified field.}

\item{keep_geometry}{logical. Should the geographical information be kept
or removed? (Defaults to TRUE).}

\item{uidlist}{A data.frame as returned by `get_uid_list()`.}
}
\value{
A simplified list with all matching docs.
}
\description{
Queries resources from an iDAI.field database that is currently running
}
\examples{
\dontrun{
idf_index_query(connection, project = "milet", field = "type", value = "Brick")
}

}
