% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{idf_ping}
\alias{idf_ping}
\title{Ping the Field-Database}
\usage{
idf_ping(idaifield_connection)
}
\arguments{
\item{idaifield_connection}{The connection as returned by `connect_idaifield()`}

\item{fail}{Should the function stop execution?}
}
\value{
returns no value
}
\description{
Helper for all functions that access the database to display
specific warning messages.
}
\examples{
\dontrun{
idaifield_connection <- connect_idaifield()
idf_ping(idaifield_connection)
}

}
\keyword{internal}
