% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_idaifield.R
\name{idf_ping}
\alias{idf_ping}
\title{Ping the Field-Database}
\usage{
idf_ping(conn)
}
\arguments{
\item{conn}{An object that contains the connection settings, as
returned by `connect_idaifield()` or a `crul::HttpClient` object
as returned by `proj_idf_client()` (internal).}
}
\value{
A boolean value indicating if the connection
was successful (TRUE) or not (FALSE).
}
\description{
This function checks if a connection to the Field Database
can be established. It returns a boolean value indicating if the
connection was successful or not.
}
\examples{
\dontrun{
# Establish a connection to the Field Database
conn <- connect_idaifield(ping = FALSE)

# Ping the Field Database
idf_ping(conn)
}
}
