% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_projects.R
\name{idf_projects}
\alias{idf_projects}
\title{Get a vector of project names present in the database}
\usage{
idf_projects(connection)
}
\arguments{
\item{connection}{The connection settings as
returned by \code{\link{connect_idaifield}}}
}
\value{
A character vector containing the projects present in the database,
NA if connection was refused.
}
\description{
Get a vector of project names present in the database
}
\examples{
\dontrun{
 connection <- connect_idaifield(pwd = "hallo")
 idf_project_list(connection)
}
}
\references{
Field Desktop Client: \url{https://github.com/dainst/idai-field}
}
