% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_query}
\alias{idf_query}
\title{idf_query}
\usage{
idf_query(connection, projectname = "NULL", field = "type", value = "Brick")
}
\arguments{
\item{connection}{A connection object as returned by `connect_idaifield()`}

\item{projectname}{The name of the project to be queried.}

\item{field}{The resource field that should be selected for (i.e. "type" for
the type of resource (Pottery, Brick, Layer)).}

\item{value}{The value to be selected for in the specified field (i.e.
"Brick" when looking for resourced of type "Brick").}
}
\value{
An 'idaifield_docs' list
}
\description{
Queries resources from an iDAI.field database that is currently running
}
\examples{
\dontrun{
conn <- connect_idaifield(pwd = "hallo")
idf_query(conn, projectname = "rtest", field = "type", value = "Brick")
}
}
