% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idf_query.R
\name{idf_query}
\alias{idf_query}
\title{idf_query}
\usage{
idf_query(
  connection,
  projectname = "NULL",
  field = "type",
  value = "Brick",
  uidlist = NULL,
  keep_geometry = TRUE
)
}
\arguments{
\item{connection}{A connection object as returned by `connect_idaifield()`}

\item{projectname}{The name of the project to be queried.}

\item{field}{The resource field that should be selected for (i.e. "type" for
the type of resource (Pottery, Brick, Layer)).}

\item{value}{The value to be selected for in the specified field (i.e.
"Brick" when looking for resourced of type "Brick").}

\item{uidlist}{A data.frame as returned by `get_uid_list()`.}

\item{keep_geometry}{logical. Should the geographical information be kept
or removed? (Defaults to TRUE).}
}
\value{
A simplified list with all matching resources.
}
\description{
Queries resources from an iDAI.field database that is currently running
}
\examples{
\dontrun{
idf_query(connection, projectname = "rtest", field = "type", value = "Brick")
}

}
