% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{idf_sepdim}
\alias{idf_sepdim}
\title{Convert a List of Dimensions to Simpler Values}
\usage{
idf_sepdim(dimensionList, name = NULL)
}
\arguments{
\item{dimensionList}{A list of dimensions (e.g. \code{dimensionLength},
\code{dimensionWidth}, etc.) from a single resource (element) of an
\code{idaifield_docs} or \code{idaifield_resources} object.}

\item{name}{character. The name of the corresponding dimension list.
Defaults to "dimension".}
}
\value{
A list containing simple values for each measured dimension
from the list. If a range was entered, the function returns the mean
without further comment.
}
\description{
This function breaks down a list of dimensions
(e.g. \code{dimensionLength}, \code{dimensionWidth}, etc.) from a single
resource (element) of an \code{idaifield_docs} or \code{idaifield_resources} object
and converts them to simple values.It is used as a helper function
for \code{simplify_idaifield()}.
}
\examples{
\dontrun{
dimensionLength_new <- idf_sepdim(idaifield_resources[[1]]$dimensionLength, "dimensionLength")
}
}
\seealso{
\itemize{
\item This function is used by: \code{\link[=simplify_idaifield]{simplify_idaifield()}}
}
}
