% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_idaifield.R
\name{idf_show_categories}
\alias{idf_show_categories}
\title{Show the categories present in an `idaifield_docs`- or
`idaifield_resources`-list}
\usage{
idf_show_categories(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An an `idaifield_docs`- or `idaifield_resources`-list
as returned by \code{\link{get_idaifield_docs}} or \code{\link{idf_query}}
and \code{\link{idf_index_query}}.}
}
\value{
A character vector with the unique categories present in
the iDAI.field 2 / Field Desktop database the list was imported from.
}
\description{
Returns a list of all categories present in the
iDAI.field 2 / Field Desktop database the list was imported from.
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
                                user = "R",
                                pwd = "hallo",
                                project = "rtest")

idaifield_docs <- get_idaifield_docs(connection = connection)

idf_show_categories(idaifield_docs)
}

}
\references{
Field Desktop Client: \url{https://github.com/dainst/idai-field}
}
