% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_if_empty}
\alias{na_if_empty}
\title{na_if_empty: returns NA if an object handed to the function is empty}
\usage{
na_if_empty(item)
}
\arguments{
\item{item}{any object whatsoever}
}
\value{

}
\description{
This is a helper function in defense against empty list items from `idaifield`,
which sometimes occur. It simply writes NA in the corresponding field if a
list or any kind of object handed to it is of length 0.
Otherwise, it returns the input untouched.
}
