% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_idaifield.R
\name{proj_idf_client}
\alias{proj_idf_client}
\title{Connection-Client to a Field Database to get all docs (internal)}
\usage{
proj_idf_client(conn, project = NULL, include = "all")
}
\arguments{
\item{conn}{The connection settings as returned by `connect_idaifield()`}

\item{project}{character. Name of the project-database that should be loaded.}
}
\value{
a crul::HttpClient
}
\description{
Connection-Client to a Field Database to get all docs (internal)
}
\examples{
\dontrun{
  connection <- connect_idaifield(pwd = "hallo", project = "rtest")
  client <- proj_idf_client(conn = connection)
}
}
\keyword{internal}
