% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_geometry.R
\name{reformat_geometry}
\alias{reformat_geometry}
\title{Reformat the Geometry of a single iDAI.field / Field Desktop resource}
\usage{
reformat_geometry(geometry)
}
\arguments{
\item{geometry}{The list in \code{resource$geometry} of an \code{idaifield_docs}-
or \code{idaifield_resources}-list.}
}
\value{
The geometry of the resource in a more usable format that can
be processed with tools like the \code{\link[sp:SpatialPoints]{sp::SpatialPoints()}} etc. to work with
spatial data.
}
\description{
Reformat the Geometry of a single iDAI.field / Field Desktop resource
}
\examples{
\dontrun{
test_2 <- list(type = "Polygon",
    coordinates = list(list(list(1), list(1)),
                       list(list(1), list(2)),
                       list(list(2), list(2)),
                       list(list(1), list(1))
                       )
)

reformat_geometry(test_2)

}
}
\seealso{
\itemize{
\item This function is used in: \code{\link[=simplify_idaifield]{simplify_idaifield()}}
}
}
