% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{remove_config_names}
\alias{remove_config_names}
\title{Remove everything before the : in a character vector}
\usage{
remove_config_names(nameslist = c("identifier", "configname:test"))
}
\arguments{
\item{nameslist}{a character vector}
}
\value{
same character vector without everything before
the ":" including the ":"
}
\description{
This function is a helper to `simplify_single_resource()`.
}
\examples{
\dontrun{
nameslist <- c("relation.liesWithin","relation.liesWithinLayer",
"campaign.2022","milet:test")
nameslist <- remove_config_names(nameslist)
nameslist
}
}
\keyword{internal}
