% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{remove_config_names}
\alias{remove_config_names}
\title{Remove Everything Before the Colon in a Character Vector}
\usage{
remove_config_names(
  conf_names = c("identifier", "configname:test"),
  silent = FALSE
)
}
\arguments{
\item{conf_names}{A character vector.}

\item{silent}{Should the message that duplicates were
detected be suppressed? Default is FALSE.}
}
\value{
The same character vector with everything before
the first colon (including the colon) removed.
}
\description{
This function removes everything before the first
colon (including the colon) in a character vector.
It is used as a helper function for \code{simplify_idaifield()}.
}
\examples{
\dontrun{
nameslist <- c("relation.liesWithin", "relation.liesWithinLayer",
"campaign.2022", "rtest:test", "pergamon:neuesFeld")
remove_config_names(nameslist, silent = FALSE)
}
}
\seealso{
\itemize{
\item This function is used by: \code{\link[=simplify_idaifield]{simplify_idaifield()}}, \code{\link[=get_field_index]{get_field_index()}}
and \code{\link[=get_uid_list]{get_uid_list()}}
}
}
