% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_utilities.R
\name{remove_config_names}
\alias{remove_config_names}
\title{Remove everything before the colon in a character vector}
\usage{
remove_config_names(conf_names = c("identifier", "configname:test"))
}
\arguments{
\item{conf_names}{A character vector.}
}
\value{
The same character vector with everything before
the first colon (including the colon) removed.
}
\description{
This function removes everything before the first
colon (including the colon) in a character vector.
It is used as a helper function for `simplify_idaifield()`.
}
\examples{
\dontrun{
nameslist <- c("relation.liesWithin", "relation.liesWithinLayer",
"campaign.2022", "rtest:test", "pergamon:neuesFeld")
remove_config_names(nameslist)
}
}
\seealso{
\code{\link{simplify_idaifield}}
}
