% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_uid.R
\name{replace_uid}
\alias{replace_uid}
\title{Replace a vector of UUIDs with their identifiers}
\usage{
replace_uid(uidvector, uidlist)
}
\arguments{
\item{uidvector}{a vector of UUIDs to be replaced with their identifiers}

\item{uidlist}{a uidlist resp. index of as returned by
\code{\link{get_uid_list}} and \code{\link{get_field_index}}}
}
\value{
The corresponding identifier(s) (a character string/vector)
}
\description{
When handed an item (a vector or a single variable) first checks if it is
actually a UUID as defined in check_if_uid() and if so, replaces it with
the corresponding identifier from the uidlist (also handed to the function).
}
\examples{
\dontrun{
uidlist <- get_uid_list(idaifield_docs)

replace_uid("9e436b96-134d-e610-c032-136fc9e8e26e", uidlist = uidlist)
}
}
\seealso{
\code{\link{simplify_idaifield}}, \code{\link{fix_relations}}
}
