% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_uid.R
\name{replace_uid}
\alias{replace_uid}
\title{replace_uid}
\usage{
replace_uid(uidvector, uidlist)
}
\arguments{
\item{uidvector}{a vector of UIDs to be replaced with their identifiers (if
numeric, the corresponging column of the uidlist is processed)}

\item{uidlist}{a uidlist as returned by get_uid_list()}
}
\value{
The corresponding identifier(s) (a character string/vector)
}
\description{
When handed an item (a vector or a single variable) first checks if it is
actually a UID as defined in check_if_uid() and if so, replaces it with
the corresponding identifier from the uidlist (also handed to the function).
}
\examples{
\dontrun{
uidlist <- get_uid_list(idaifield_docs)

replace_uid("9e436b96-134d-e610-c032-136fc9e8e26e", uidlist = uidlist)
}
}
\keyword{internal}
