% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_idaifield.R
\name{select_by}
\alias{select_by}
\title{select_by}
\usage{
select_by(idaifield_docs, by = c("type", "isRecordedIn"), value = NULL)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs(...)}

\item{by}{must be either type (to select by resource type) or isRecordedIn
(to select by container-resource (Survey-Area, Trench))}

\item{value}{Character expected, should be the internal Name of the Type
that will be selected for (e.g. "Layer", "Pottery"), can be vector of
multiple Types / Operations}
}
\value{
a list of class idaifield_resources containing the resources
which are of the selected type
}
\description{
returns a subset of the docs list selected by type or isRecordedIn
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
idaifield_docs <- get_idaifield_docs(connection = connection,
projectname = "rtest")

idaifield_layers <- select_by_type(idaifield_docs,
by = "type",
value = "Layer")
}
}
