% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_idaifield.R
\name{select_by}
\alias{select_by}
\title{select_by}
\usage{
select_by(idaifield_docs, by = c("category", "isRecordedIn"), value = NULL)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs(...)}

\item{by}{must be either type (to select by resource category) or
isRecordedIn (to select by container-resource (Survey-Area, Trench))}

\item{value}{Character expected, should be the internal Name of the
that will be selected for (e.g. "Layer", "Pottery"), can be vector of
multiple Categories / Operations}
}
\value{
a list of class idaifield_resources containing the resources
which are of the selected category
}
\description{
returns a subset of the docs list selected by category or isRecordedIn
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo", project = "rtest")
idaifield_docs <- get_idaifield_docs(connection = connection)

idaifield_layers <- select_by(idaifield_docs,
by = "category",
value = "Layer")
}
}
