% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importlist.R
\name{select_by_type}
\alias{select_by_type}
\title{select_by_type}
\usage{
select_by_type(idaifield_docs, type = "Pottery")
}
\arguments{
\item{type}{}
}
\value{

}
\description{
returns a subset of the docs list selected by type
}
