% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_idaifield.R
\name{show_type_list}
\alias{show_type_list}
\title{show_type_list}
\usage{
show_type_list(idaifield_docs)
}
\arguments{
\item{idaifield_docs}{An object as returned by get_idaifield_docs(...);
unnests to resource level if it didn't already happen.}
}
\value{
a character vector with the types represented in
the idaifield_docs/resource
}
\description{
Returns a list of types present in the iDAI.field 2-database
(for orientation) it would be nice to be able to embed the translations,
but i am guessing thats only possible if the configuration was available
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "192.168.2.21",
user = "R", pwd = "hallo")
idaifield_docs <- get_idaifield_docs(connection = connection,
projectname = "rtest")

show_type_list(idaifield_docs)
}
}
