% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_idaifield}
\alias{simplify_idaifield}
\title{Title}
\usage{
simplify_idaifield(idaifield_docs, replace_uids = TRUE, uidlist = "generate")
}
\arguments{
\item{uidlist}{#'}

\item{keep_geometry}{FALSE by default. TRUE if you wish to import the
geometry-fields into R (TODO: this is not reachable from top level function)

needs: keep_geometry; unnest relations}
}
\value{

}
\description{
Title
}
