% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_idaifield}
\alias{simplify_idaifield}
\title{Simplify a list imported from an iDAI.field-Database}
\usage{
simplify_idaifield(
  idaifield_docs,
  keep_geometry = TRUE,
  replace_uids = TRUE,
  uidlist = NULL,
  language = "all"
)
}
\arguments{
\item{idaifield_docs}{An "idaifield_docs" or "idaifield_resources"-list as
returned by `get_idaifield_docs()`.}

\item{keep_geometry}{logical. Should the geographical information be kept
or removed? (Defaults to TRUE).}

\item{replace_uids}{logical. Should UIDs be automatically replaced with the
corresponding identifiers? (Defaults to TRUE).}

\item{uidlist}{If NULL (default) the list of UIDs and identifiers is
automatically generated within this function. This only makes sense if
the list handed to `simplify_idaifield()` had not been selected yet. If it
has been, you should supply a data.frame as returned by `get_uid_list()`.}

\item{language}{the short name (e.g. "en", "de", "fr") of the language that
is preferred for the fields, defaults to english ("en")}
}
\value{
a simplified "idaifield_resources"-list
}
\description{
Simplify a list imported from an iDAI.field-Database
}
\examples{
\dontrun{
connection <- connect_idaifield(serverip = "127.0.0.1",
user = "R", pwd = "hallo")
idaifield_docs <- get_idaifield_docs(connection = connection,
projectname = "rtest")

simpler_idaifield <- simplify_idaifield(idaifield_docs)
}
}
