% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_single_resource}
\alias{simplify_single_resource}
\title{Simplifies a single resource from the iDAI.field 2 / Field Desktop Database}
\usage{
simplify_single_resource(
  resource,
  replace_uids = TRUE,
  uidlist = NULL,
  keep_geometry = TRUE,
  fieldtypes = NULL,
  language = "all",
  spread_fields = TRUE,
  use_exact_dates = FALSE
)
}
\arguments{
\item{resource}{One resource (element) from an idaifield_resources-list.}

\item{replace_uids}{see `?simplify_idaifield()`}

\item{uidlist}{see `?simplify_idaifield()`}

\item{keep_geometry}{see `?simplify_idaifield()`}

\item{language}{see `?simplify_idaifield()`}

\item{spread_fields}{see `?simplify_idaifield()`}
}
\value{
A single resource (element) for an idaifield_resource-list.
}
\description{
This function is a helper to `simplify_idaifield()`.
}
\examples{
\dontrun{
simpler_resource <- simplify_single_resource(resource,
replace_uids = TRUE,
uidlist = uidlist,
keep_geometry = FALSE)
}
}
\keyword{internal}
