% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_single_resource}
\alias{simplify_single_resource}
\title{Title}
\usage{
simplify_single_resource(
  resource,
  replace_uids = TRUE,
  uidlist = uidlist,
  keep_geometry = FALSE
)
}
\arguments{
\item{keep_geometry}{}
}
\value{

}
\description{
Title
}
