% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_idaifield_resources.R
\name{simplify_single_resource}
\alias{simplify_single_resource}
\title{Simplifies a single resource from the iDAI.field 2 / Field Desktop Database}
\usage{
simplify_single_resource(
  resource,
  replace_uids = TRUE,
  find_layers = TRUE,
  uidlist = NULL,
  keep_geometry = TRUE,
  fieldtypes = NULL,
  language = "all",
  spread_fields = TRUE,
  use_exact_dates = FALSE
)
}
\arguments{
\item{resource}{One resource (element) from an \code{idaifield_resources}-list.}

\item{replace_uids}{TRUE/FALSE: Should UUIDs be automatically replaced with the
corresponding identifiers? Defaults is TRUE. Uses: \code{\link[=fix_relations]{fix_relations()}} with
\code{\link[=replace_uid]{replace_uid()}}, and also: \code{\link[=find_layer]{find_layer()}}}

\item{find_layers}{TRUE/FALSE. Default is FALSE. If TRUE, adds another column
with the 'Layer' (see \code{getOption("idaifield_categories")$layers}, can be
modified) in which a resource is contained  recursively. That means that
even if it does not immediately lie within this layer, but is
contained by one or several other resources in said layer, a new column
("liesWithinLayer") will still show the layer.
Example: A sample "A" in Find "001" from layer "Layer1" will
usually have "001" as the value in "liesWithin". With find_layers, there will
be another column called "liesWithinLayer" which contains "Layer1" for both
sample "A" and Find "001".}

\item{uidlist}{If NULL (default) the list of UUIDs and identifiers is
automatically generated within this function using \code{\link[=get_uid_list]{get_uid_list()}}. This only makes sense if
the list handed to \code{\link[=simplify_idaifield]{simplify_idaifield()}} had not been selected yet. If it
has been, you should supply a data.frame as returned
by \code{\link[=get_field_index]{get_field_index()}}.}

\item{keep_geometry}{TRUE/FALSE: Should the geographical
information be kept or removed? Defaults is FALSE. Uses: \code{\link[=reformat_geometry]{reformat_geometry()}}}

\item{language}{the short name (e.g. "en", "de", "fr") of the language that
is preferred for the multi-language input fields, defaults to keeping all
languages as sub-lists ("all"). Uses: \code{\link[=gather_languages]{gather_languages()}}}

\item{spread_fields}{TRUE/FALSE: Should checkbox-fields be
spread across multiple lists to facilitate boolean-columns for each value
of a checkbox-field? Default is TRUE. Uses: \code{\link[=get_configuration]{get_configuration()}},
\code{\link[=get_field_inputtypes]{get_field_inputtypes()}}, \code{\link[=convert_to_onehot]{convert_to_onehot()}}}

\item{use_exact_dates}{TRUE/FALSE: Should the values from any "exact"
dates be used in case there are any? Default is FALSE. Changes outcome of \code{\link[=fix_dating]{fix_dating()}}.}
}
\value{
A single resource (element) for an \code{idaifield_resources}-list.
}
\description{
This function is a helper to \code{simplify_idaifield()}.
}
\examples{
\dontrun{
simpler_resource <- simplify_single_resource(resource,
replace_uids = TRUE,
uidlist = uidlist,
keep_geometry = FALSE)
}
}
\keyword{internal}
